requirements.txt
pyArabic
python-Levenshtein
pytesseract
kraken
requests
easyocr
opencv-python
pandas
swifter
Pillow==9.5.0